+++ fairdm_geo/core/__init__.py
@@ -1,0 +1,8 @@
+"""
+Core abstract models for fairdm-geo.
+
+This app provides abstract base classes that other fairdm-geo apps inherit from.
+It should be installed first in INSTALLED_APPS.
+"""
+
+default_app_config = "fairdm_geo.core.apps.CoreConfig"