+++ fairdm_geo/rocks/__init__.py
@@ -1,0 +1,8 @@
+"""
+FairDM Geo - Rocks App
+
+This Django app provides concrete rock sample models for geological data management.
+It builds upon the abstract base models from fairdm_geo.core.
+"""
+
+default_app_config = "fairdm_geo.rocks.apps.RocksConfig"