+++ fairdm_geo/rocks/models/__init__.py
@@ -1,0 +1,13 @@
+"""Rock sample models."""
+
+from .base import BaseRock
+from .samples import DrillCore, DrillCuttings, RockPowder, RockSample, ThinSection
+
+__all__ = [
+    "BaseRock",
+    "RockSample",
+    "DrillCore",
+    "DrillCuttings",
+    "ThinSection",
+    "RockPowder",
+]