+++ fairdm_geo/sites/__init__.py
@@ -1,0 +1,8 @@
+"""
+FairDM Geo - Sites App
+
+This Django app provides concrete sampling location and borehole models for geographical
+sampling features. It builds upon the abstract base models from fairdm_geo.core.
+"""
+
+default_app_config = "fairdm_geo.sites.apps.SitesConfig"