+++ fairdm_geo/sites/models/__init__.py
@@ -1,3 +1,5 @@
+"""Sampling location models."""
+
 from .sites import Borehole, SamplingLocation
 
 __all__ = [